+++ marvel_reader/app/controllers/api/v1/characters_controller.rb
@@ -4,22 +4,16 @@ module Api::V1
     def index
       logger.debug("Calling service ...")
 
-      _params = {
-          resource: 'characters',
-          search: {name: params[:term]}
-      }
+      _character_finder = Services::CharacterFinder.call(params: params, accepted_params: accepted_params)
 
-      _comic_reader = ComicReader.call(_params)
+      if _character_finder.fail?
+        flash[:error] = _character_finder.message
+      end
 
-      _page_params = accepted_params
-
-      if _comic_reader.success?
-        @characters = _comic_reader.result['data']['results'].map do |data|
-          Character.new(data)
-        end
+      if _character_finder.success?
 
         # a very, very simple serialization
-        _response_json = @characters.map do |character|
+        _response_json = _character_finder.characters.map do |character|
           {
               id: character.id,
               name: character.name,
@@ -29,7 +23,7 @@ module Api::V1
 
         build_response({characters: _response_json})
       else
-        build_response({error: _comic_reader.message}, false, -1, t('errors.record_not_found'), 404)
+        build_response({error: _character_finder.message}, false, -1, t('errors.record_not_found'), 404)
       end
 
     end
